[Fix] 채용담당자 회원가입 페이지 수정 완료
